working on project page
woked on the project page adding in all the images to the project index page.  added two images from pintrest since there was no images for saunas or spas.  added all the current image assets from google grive to the images folderunder the project page.  the files can be changed if needed.